fix gussian grid area
